update to fix config.json invalid error

+++ libs/setup.js
@@ -9,21 +9,22 @@ var os = require('os')
 var configFile = path.join(__dirname, '..', 'config.json')
 var sampleConfigFile = path.join(__dirname, '..', 'config.sample.json')
 
+
+var configExists = fs.statSync(configFile).isFile()
 function writeConfigFile (config, options) {
-  fs.readFile(configFile, function (err, data) {
-    if (options.force) {
-      fs.truncate(configFile, 0, function() {
-        fs.writeFile(configFile, JSON.stringify(config, null, 2), function (err) {
-          if (err) {
-            log.error('Error writing to', configFile, err)
-          }
-          log.debug('setup.writeConfigFile -', 'Config succesfully written to', configFile)
-        })
+  var configExists = fs.statSync(configFile).isFile()
+  if (configExists && !options.force) {
+    log.error('setup.writeConfigFile', 'Error writing to', configFile, 'file exists! Pass -f flag to overwrite.')
+  } else {
+    fs.truncate(configFile, 0, function() {
+      fs.writeFile(configFile, JSON.stringify(config, null, 2), function (err) {
+        if (err) {
+          log.error('Error writing to', configFile, err)
+        }
+        log.debug('setup.writeConfigFile -', 'Config succesfully written to', configFile)
       })
-    } else {
-      log.error('setup.writeConfigFile', 'Error writing to', configFile, 'file exists! Pass -f flag to overwrite.')
-    }
-  })
+    })
+  }
 }
 
 var Setup;
@@ -41,6 +42,12 @@ Setup = function(env) {
   }
 
   var questions = [
+    {
+      type: 'input',
+      name: 'baseUrl',
+      message: 'What is the base url',
+      default: config.baseUrl
+    },
     {
       type: 'input',
       name: 'agentUrl',
@@ -139,6 +146,7 @@ Setup = function(env) {
       }
     }
   ]
+
   inquirer.prompt(questions)
   .then(function (answers) {
     config = _.merge({}, config, answers)
